beefy: use sc-utils generic notification channel
