+++ client/utils/Cargo.toml
@@ -12,9 +12,13 @@ readme = "README.md"
 [dependencies]
 futures = "0.3.9"
 lazy_static = "1.4.0"
+parking_lot = "0.11"
 prometheus = { version = "0.13.0", default-features = false }
 futures-timer = "3.0.2"
 
 [features]
 default = ["metered"]
 metered = []
+
+[dev-dependencies]
+tokio-test = "0.4.2"
